# src/docs/swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SellMate API",
    "version": "1.0.0",
    "description": "\n# SellMate Platform - Social Commerce Automation Backend API\n\nSellMate is an intelligent social commerce automation platform designed specifically for Nigerian micro-entrepreneurs who sell through Instagram and WhatsApp.\n\n## Features\n- **Unified Conversations**: Consolidate WhatsApp and Instagram messages in one place\n- **AI-Powered Order Detection**: Automatically detect orders from conversations\n- **Profit Analytics**: Real-time business metrics and profit calculations\n- **Customer Management**: Track customer interactions across platforms\n\n## Authentication\nMost endpoints require JWT authentication. Include the token in the Authorization header:\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n## Rate Limiting\nAPI requests are rate-limited to prevent abuse. Current limits:\n- General: 100 requests per 15 minutes\n- Authentication: 5 requests per 15 minutes\n- AI endpoints: 20 requests per minute\n      ",
    "contact": {
      "name": "SellMate Team",
      "email": "support@sellmate.ng"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.sellmate.ng",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT access token obtained from /auth/verify-otp or /auth/refresh"
      }
    },
...
  },
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/analytics/customers": {
      "get": {
        "summary": "Get customer analytics",
        "tags": [
          "Analytics"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "startDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "in": "query",
            "name": "endDate",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer analytics with segmentation and top customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "totalCustomers": {
                          "type": "number"
                        },
                        "newCustomers": {
                          "type": "number"
                        },
                        "returningCustomers": {
                          "type": "number"
                        },
                        "topCustomers": {
                          "type": "array"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/conversations/{id}/customer": {
      "put": {
        "summary": "Link customer to conversation",
        "tags": [
          "Conversations"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customerId"
                ],
                "properties": {
                  "customerId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer linked"
          },
          "404": {
            "description": "Conversation not found"
          }
        }
      }
    },
    "/api/v1/customers": {
      "get": {
        "summary": "List customers for authenticated user",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "platform",
            "schema": {
              "type": "string",
              "enum": [
                "whatsapp",
                "instagram"
              ]
            },
            "description": "Filter by platform"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by name or phone"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of customers"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create a new customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "platforms": {
                    "type": "object",
                    "properties": {
                      "whatsapp": {
                        "type": "object",
                        "properties": {
                          "phoneNumber": {
                            "type": "string"
                          },
                          "profileName": {
                            "type": "string"
                          }
                        }
                      },
                      "instagram": {
                        "type": "object",
                        "properties": {
                          "username": {
                            "type": "string"
                          },
                          "profileName": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "summary": "Get customer details",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "put": {
        "summary": "Update customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string"
                  },
                  "notes": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      },
      "delete": {
        "summary": "Delete customer",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Customer deleted"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
    "/api/v1/customers/{id}/insights": {
      "get": {
        "summary": "Get customer insights",
        "tags": [
          "Customers"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer insights including order history and platform info"
          },
          "404": {
            "description": "Customer not found"
          }
        }
      }
    },
...
    "/api/v1/orders/customer/{customerId}": {
      "get": {
        "summary": "Get orders by customer",
        "tags": [
          "Orders"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "customerId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer orders"
          }
        }
      }
    },
...
  }
}